renamed iconLink to just link
