+++ src/components/bs5/link/link.mdx
@@ -1,12 +1,12 @@
 import { Canvas, Meta, Story } from "@storybook/blocks";
-import * as iconLinkStories from "./iconLink.stories";
+import * as linkStories from "./link.stories";
 
-<Meta of={iconLinkStories} />
+<Meta of={linkStories} />
 
 # Icon Link
 
 <Canvas>
-  <Story of={iconLinkStories.Default} />
+  <Story of={linkStories.Default} />
 </Canvas>
 
 ## Design resources
