+++ src/css/main.scss
@@ -110,7 +110,7 @@ $enable-dark-mode: true;
 @import "../components/bs5/formcheck/formcheck";
 @import "../components/bs5/footer/footer";
 @import "../components/bs5/globalAlert/globalAlert";
-@import "../components/bs5/iconLink/iconLink";
+@import "../components/bs5/link/link";
 @import "../components/bs5/header/header";
 @import "../components/bs5/icons/icons";
 @import "../components/bs5/inpageAlert/inpageAlert";
