+++ src/js/handlebars.partials.js
@@ -8,8 +8,8 @@ import blockquote from "../components/bs5/blockquote/blockquote.hbs?raw";
 import breadcrumbs from "../components/bs5/breadcrumbs/breadcrumbs.hbs?raw";
 import breadcrumbsWrapper from "../components/bs5/breadcumbsWrapper/breadcrumbsWrapper.hbs?raw";
 import button from "../components/bs5/button/button.hbs?raw";
-import callToAction from "../components/bs5/callToAction/callToAction.hbs?raw";
 import callout from "../components/bs5/callout/callout.hbs?raw";
+import callToAction from "../components/bs5/callToAction/callToAction.hbs?raw";
 import card from "../components/bs5/card/card.hbs?raw";
 import containerLayout from "../components/bs5/containerLayout/containerLayout.hbs?raw";
 import contentFooter from "../components/bs5/contentFooter/contentFooter.hbs?raw";
@@ -25,11 +25,11 @@ import fullPage from "../components/bs5/fullPageWrapper/fullPage.hbs?raw";
 import globalAlert from "../components/bs5/globalAlert/globalAlert.hbs?raw";
 import head from "../components/bs5/head/head.hbs?raw";
 import header from "../components/bs5/header/header.hbs?raw";
-import iconLink from "../components/bs5/iconLink/iconLink.hbs?raw";
 import icon from "../components/bs5/icons/icon.hbs?raw";
 import image from "../components/bs5/image/image.hbs?raw";
 import inpageAlert from "../components/bs5/inpageAlert/inpageAlert.hbs?raw";
 import inpagenav from "../components/bs5/inpagenav/inpagenav.hbs?raw";
+import link from "../components/bs5/link/link.hbs?raw";
 import linkColumns from "../components/bs5/linkColumns/linkColumns.hbs?raw";
 import logo from "../components/bs5/logo/logo.hbs?raw";
 import logoCOALandscape from "../components/bs5/logo/logoCOALandscape.hbs?raw";
@@ -74,8 +74,8 @@ export default function handlebarsPartials(handlebars) {
   handlebars.registerPartial("breadcrumbs", breadcrumbs);
   handlebars.registerPartial("breadcrumbsWrapper", breadcrumbsWrapper);
   handlebars.registerPartial("button", button);
-  handlebars.registerPartial("callToAction", callToAction);
   handlebars.registerPartial("callout", callout);
+  handlebars.registerPartial("callToAction", callToAction);
   handlebars.registerPartial("card", card);
   handlebars.registerPartial("containerLayout", containerLayout);
   handlebars.registerPartial("contentFooter", contentFooter);
@@ -91,11 +91,11 @@ export default function handlebarsPartials(handlebars) {
   handlebars.registerPartial("globalAlert", globalAlert);
   handlebars.registerPartial("head", head);
   handlebars.registerPartial("header", header);
-  handlebars.registerPartial("iconLink", iconLink);
   handlebars.registerPartial("icon", icon);
   handlebars.registerPartial("image", image);
   handlebars.registerPartial("inpageAlert", inpageAlert);
   handlebars.registerPartial("inpagenav", inpagenav);
+  handlebars.registerPartial("link", link);
   handlebars.registerPartial("linkColumns", linkColumns);
   handlebars.registerPartial("logo", logo);
   handlebars.registerPartial("logoCOALandscape", logoCOALandscape);
